pyproject.toml
[project]
name = "talkbank_utils"
version = "0.1.0"
description = "Utilities for talkbank transcription folders focused on text and audio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyannote-audio>=4.0.0",
    "regex>=2025.9.18",
]
[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pip>=25.2",
]
